pyproject.toml
[project]
name = "lahella-cli"
version = "0.1.0"
description = "CLI tool for managing Lähellä.fi activity listings via YAML"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Jouni Seppänen", email = "jks@iki.fi"}
]
requires-python = ">=3.14"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "filelock>=3.16.0",
    "httpx>=0.28.1",
    "playwright>=1.57.0",
    "ruamel-yaml>=0.18.17",
]

[project.urls]
Homepage = "https://github.com/jkseppan/lahella-cli"
Repository = "https://github.com/jkseppan/lahella-cli"

[dependency-groups]
dev = [
    "coverage>=7.13.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.14.10",
    "ty>=0.0.8",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "TRY",    # tryceratops (exception handling)
    "RUF",    # Ruff-specific rules
]
ignore = [
    "TRY003", # Avoid specifying long messages outside the exception class
    "TRY300", # Consider moving this statement to an else block
]
allowed-confusables = ["–"]

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build]
exclude = ["*.lock"]

[tool.ty.src]
# Exclude test files - type errors there would just cause test failures anyway
exclude = ["tests/test_*.py"]
